refactor: [판매자용] UserRepository 위치 변경
- repository 패키지 하위로 이동
